Aplicação a funcionar pela 1a vez

+++ app/src/main/java/com/viajar/viajar/database/Location.java
@@ -13,7 +13,7 @@ public class Location {
 
     public double latitude;
     public double longitude;
-    public double altitude;
+    public int altitude;
 
     @ColumnInfo(name = "extra_info")
     public String extraInfo;
